ANA-964: FX Fwd lifecycle refactoring
