+++ sdk/Lusid.Sdk.Tests/tutorials/Ibor/PortfolioCashFlows.cs
@@ -219,120 +219,7 @@ namespace Lusid.Sdk.Tests.tutorials.Ibor
 
             _recipeApi.DeleteConfigurationRecipe(_portfolioScope, modelRecipeCode);
         }
-        
-        [Test]
-        public void LifeCycleManagementForFxForward()
-        {
-            // CREATE FX Forward
-            var fxForward = (FxForward) InstrumentExamples.CreateExampleFxForward(isNdf: false);
-            
-            // CREATE wide enough window to pick up all cashflows for the FX Forward
-            var windowStart = fxForward.StartDate.AddMonths(-1);
-            var windowEnd = fxForward.MaturityDate.AddMonths(1);
-        
-            // UPSERT FX Forward to portfolio and populating stores with required market data - use a constant FX rate USD/JPY = 150.
-            var effectiveAt = windowStart;
-            AddInstrumentsTransactionPortfolioAndPopulateRequiredMarketData(
-                _portfolioScope, 
-                _portfolioCode,
-                windowStart,
-                windowEnd,
-                new List<LusidInstrument>(){fxForward},
-                useConstantFxRate: true);
 
-            // CREATE and upsert CTVoM recipe specifying discount pricing model
-            var modelRecipeCode = "CTVoMRecipe";
-            CreateAndUpsertRecipe(
-                modelRecipeCode,
-                _portfolioScope,
-                ModelSelection.ModelEnum.ConstantTimeValueOfMoney,
-                windowValuationOnInstrumentStartEnd: true);
-
-            // GET all upsertable cashflows (transactions) for the FX Forward
-            var allFxFwdCashFlows = _transactionPortfoliosApi.GetUpsertablePortfolioCashFlows(
-                _portfolioScope,
-                _portfolioCode,
-                effectiveAt,
-                windowStart,
-                windowEnd,
-                null,
-                null,
-                _portfolioScope,
-                modelRecipeCode)
-                .Values;
-
-            // There are exactly two cashflows associated to FX forward (one in each currency) both at maturity.
-            Assert.That(allFxFwdCashFlows.Count, Is.EqualTo(2));
-            Assert.That(allFxFwdCashFlows.First().TotalConsideration.Currency, Is.EqualTo("USD"));
-            Assert.That(allFxFwdCashFlows.Last().TotalConsideration.Currency, Is.EqualTo("JPY"));
-            Assert.That(allFxFwdCashFlows.Select(c => c.TransactionDate).Distinct().Count(), Is.EqualTo(1));
-            var cashFlowDate = allFxFwdCashFlows.First().TransactionDate;
-            
-            // CREATE valuation schedule 2 days before, day of and 2 days after the cashflows. 
-            var valuationSchedule = new ValuationSchedule(
-                effectiveAt: cashFlowDate.AddDays(2).ToString("o"),
-                effectiveFrom: cashFlowDate.AddDays(-2).ToString("o"));
-            
-            // CREATE valuation request for this FX Forward portfolio,
-            // where the valuation schedule covers before, at and after the expiration of the FX Forward. 
-            var valuationRequest = new ValuationRequest(
-                new ResourceId(_portfolioScope, modelRecipeCode),
-                portfolioEntityIds: new List<PortfolioEntityId> {new PortfolioEntityId(_portfolioScope, _portfolioCode)},
-                valuationSchedule: valuationSchedule,
-                metrics: ValuationSpec,
-                groupBy:  null,
-                sort:  null,
-                asAt: null,
-                reportCurrency: "USD");
-            
-            // CALL GetValuation before upserting back the cashflows. We check that when the FX Forward has expired, the PV is zero.
-            var valuationBeforeAndAfterExpirationOfFxForward = _aggregationApi.GetValuation(valuationRequest).Data;
-            foreach (var valuationResult in valuationBeforeAndAfterExpirationOfFxForward)
-            {
-                var date = (DateTime) valuationResult[ValuationDateKey];
-                var fxForwardPv = (double) valuationResult[ValuationPv];
-                if (date < fxForward.MaturityDate)
-                {
-                    Assert.That(fxForwardPv, Is.Not.EqualTo(0).Within(1e-12));
-                }
-                else
-                {
-                    Assert.That(fxForwardPv, Is.EqualTo(0).Within(1e-12));
-                }
-            }
-
-            // UPSERT the cashflows back into LUSID. We first populate the cashflow transactions with unique IDs.
-            var upsertCashFlowTransactions = PopulateCashFlowTransactionWithUniqueIds(allFxFwdCashFlows, fxForward.DomCcy);
-            _transactionPortfoliosApi.UpsertTransactions(_portfolioScope, _portfolioCode, MapToCashFlowTransactionRequest(upsertCashFlowTransactions));
-            
-            // HAVING upserted cashflow into lusid, we call GetValuation again.
-            var valuationAfterUpsertingCashFlows = _aggregationApi.GetValuation(valuationRequest).Data;
-            
-            // ASSERT portfolio PV is constant across time by grouping the valuation result by date.
-            // (constant because we upserted the cashflows back in with ConstantTimeValueOfMoney model and the FX rate is constant)
-            // That is, we are checking instrument PV + cashflow PV = constant both before and after maturity  
-            var resultsGroupedByDate = valuationAfterUpsertingCashFlows
-                .GroupBy(d => (DateTime) d[ValuationDateKey]);
-            
-            // CONVERT and AGGREGATE all results to USD
-            var pvsInUsd = resultsGroupedByDate
-                .Select(pvGroup => pvGroup.Sum(record =>
-                {
-                    var fxRate = ((string) record[ValuationCcy]).Equals("JPY") ? 1m/150 : 1;
-                    return Convert.ToDecimal(record[ValuationPv]) * fxRate;
-                }));
-
-            // ASSERT portfolio PV is constant over time within a tolerance
-            var uniquePvsAcrossDates = pvsInUsd
-                .Select(pv => Math.Round(pv, 12))
-                .Distinct()
-                .ToList();
-            Assert.That(uniquePvsAcrossDates.Count, Is.EqualTo(1));
-            
-            // CLEAN up.
-            _recipeApi.DeleteConfigurationRecipe(_portfolioScope, modelRecipeCode);
-        }
-        
         // This method maps a list of Transactions to a list of TransactionRequests so that they can be upserted back into LUSID.
         internal static List<TransactionRequest> MapToCashFlowTransactionRequest(IEnumerable<Transaction> transactions)
         {
